+++ src/systems/dnd5e-2.0.3.js
@@ -1,6 +1,8 @@
 export default {
 
-  "VERSION": "1.0.3",
+  "SYSTEM_VERSION": "2.1.0",
+
+  "VERSION": "1.0.2",
 
   // The actor class type is the type of actor that will be used for the default item pile actor that is created on first item drop.
   "ACTOR_CLASS_TYPE": "character",
@@ -9,7 +11,7 @@ export default {
   "ITEM_QUANTITY_ATTRIBUTE": "system.quantity",
 
   // The item price attribute is the path to the attribute on each item that determine how much it costs
-  "ITEM_PRICE_ATTRIBUTE": "system.price.value",
+  "ITEM_PRICE_ATTRIBUTE": "system.price",
 
   // Item filters actively remove items from the item pile inventory UI that users cannot loot, such as spells, feats, and classes
   "ITEM_FILTERS": [
@@ -39,21 +41,6 @@ export default {
       }
     }
     return itemData;
-  },
-
-  // This function is an optional system handler that specifically transforms an item's price into a more unified numeric format
-  "ITEM_COST_TRANSFORMER": (item, currencies) => {
-    const overallCost = Number(getProperty(item, "system.price.value")) ?? 0;
-    const priceDenomination = getProperty(item, "system.price.denomination");
-    if (priceDenomination) {
-      const currencyDenomination = currencies.find(currency => {
-        return currency.abbreviation.toLowerCase().includes(priceDenomination);
-      });
-      if (currencyDenomination) {
-        return overallCost * currencyDenomination.exchangeRate;
-      }
-    }
-    return overallCost;
   },
 
   // Item similarities determines how item piles detect similarities and differences in the system